Abstracted SolidityCallResultImpl from Standalone
